Started migrating definitions editor to separate class.
