+++ src/main/java/net/krazyweb/cataclysm/mapeditor/map/MapManager.java
@@ -4,12 +4,10 @@ import com.google.common.eventbus.EventBus;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.geometry.HPos;
-import javafx.geometry.Insets;
 import javafx.geometry.VPos;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.TextFieldTreeCell;
 import javafx.scene.layout.*;
-import javafx.stage.Modality;
 import net.krazyweb.cataclysm.mapeditor.ApplicationSettings;
 import net.krazyweb.cataclysm.mapeditor.MapRenderer;
 import net.krazyweb.cataclysm.mapeditor.events.FileLoadedEvent;
@@ -17,11 +15,9 @@ import net.krazyweb.cataclysm.mapeditor.events.FileSavedEvent;
 import net.krazyweb.cataclysm.mapeditor.events.MapChangedEvent;
 import net.krazyweb.cataclysm.mapeditor.events.UndoBufferChangedEvent;
 import net.krazyweb.cataclysm.mapeditor.map.data.*;
-import net.krazyweb.cataclysm.mapeditor.map.data.utils.PropertySheetItemCreator;
 import net.krazyweb.cataclysm.mapeditor.map.undo.UndoBufferListener;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.controlsfx.control.PropertySheet;
 
 import java.io.BufferedWriter;
 import java.io.IOException;
@@ -363,79 +359,12 @@ public class MapManager implements UndoBufferListener {
 
 	public void editDefinitions() {
 
-		//TODO Extract to own class
-		Dialog<Boolean> definitionsDialog = new Dialog<>();
-		definitionsDialog.setTitle("Edit Definitions");
-		definitionsDialog.initModality(Modality.APPLICATION_MODAL);
-		definitionsDialog.setResizable(true);
+		DefinitionsEditor definitionsEditor = new DefinitionsEditor(overMapEntries);
+		definitionsEditor.showAndWait();
 
-		SplitPane parent = new SplitPane();
-		parent.setDividerPosition(0, 0.3);
-		parent.setPadding(new Insets(0));
-
-		TreeItem<String> treeRoot = new TreeItem<>("");
-
-		TreeItem<String> itemGroups = new TreeItem<>("Item Groups");
-		itemGroups.setExpanded(true);
-		treeRoot.getChildren().add(itemGroups);
-
-		TreeItem<String> monsterGroups = new TreeItem<>("Monster Groups");
-		monsterGroups.setExpanded(true);
-		treeRoot.getChildren().add(monsterGroups);
-
-		TreeItem<String> overMaps = new TreeItem<>("Overmaps");
-		overMaps.setExpanded(true);
-		treeRoot.getChildren().add(overMaps);
-
-		overMapEntries.forEach(overMapEntry -> {
-			TreeItem<String> overMap = new TreeItem<>(overMapEntry.name);
-			overMaps.getChildren().add(overMap);
-		});
-
-		TreeView<String> treeView = new TreeView<>(treeRoot);
-		treeView.setShowRoot(false);
-		treeView.setEditable(true);
-		treeView.setCellFactory(factory -> new TreeCell());
-		treeView.getSelectionModel().selectedItemProperty().addListener((observable, oldValue, newValue) -> {
-			if (newValue.getParent() == overMaps) {
-				parent.getItems().remove(1);
-				parent.getItems().add(new PropertySheet(PropertySheetItemCreator.getPropertySheetItems(overMapEntries.get(newValue.getParent().getChildren().indexOf(newValue)))));
-			}
-		});
-
-		parent.getItems().add(treeView);
-		parent.getItems().add(new PropertySheet());
-
-		ButtonType closeButton = new ButtonType("Close", ButtonBar.ButtonData.OK_DONE);
-		definitionsDialog.getDialogPane().getButtonTypes().setAll(closeButton);
-
-		definitionsDialog.setOnCloseRequest(event -> definitionsDialog.close());
-
-		definitionsDialog.getDialogPane().setContent(parent);
-		definitionsDialog.showAndWait();
+		//TODO Get modified entries if accepted changes
 
 		updateUndoRedoText();
-
-	}
-
-	private class TreeCell extends TextFieldTreeCell {
-
-		private ContextMenu contextMenu;
-
-		public TreeCell() {
-			super();
-			this.setEditable(false);
-			contextMenu = new ContextMenu();
-			MenuItem testItem = new MenuItem("TEST");
-			contextMenu.getItems().add(testItem);
-		}
-
-		@Override
-		@SuppressWarnings("unchecked")
-		public void updateItem(Object item, boolean empty) {
-			super.updateItem(item, empty);
-			setContextMenu(contextMenu);
-		}
 
 	}
